requirements.txt
nltk
numpy
pandas
pdfplumber
pytesseract
tesseract
scikit-learn
sentence-transformers
spacy
streamlit
opencv-python-headless
Pillow
python-docx
mpmath
sympy
safetensors
pyyaml
networkx
fsspec
filelock
torch
huggingface-hub
tokenizers
transformers
